#114 Refactor boilerplate code in Search API. Repositories beans creation common logic.
