+++ search-api/src/main/kotlin/fund/cyber/api/bitcoin/TxHandlers.kt
@@ -3,7 +3,7 @@ package fund.cyber.api.bitcoin
 import fund.cyber.api.common.asSingleRouterFunction
 import fund.cyber.cassandra.bitcoin.model.CqlBitcoinTx
 import fund.cyber.cassandra.bitcoin.repository.BitcoinTxRepository
-import fund.cyber.cassandra.configuration.REPOSITORY_NAME_DELIMETER
+import fund.cyber.cassandra.common.REPOSITORY_NAME_DELIMITER
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
@@ -16,7 +16,7 @@ import org.springframework.web.reactive.function.server.RouterFunctions
 import org.springframework.web.reactive.function.server.ServerResponse
 
 @Configuration
-@DependsOn("bitcoin-cassandra-repositories")
+@DependsOn("bitcoin-search-repositories")
 class BitcoinTxHandlersConfiguration {
 
     @Autowired
@@ -26,7 +26,7 @@ class BitcoinTxHandlersConfiguration {
     fun bitcoinTxByHash(): RouterFunction<ServerResponse> {
 
         return applicationContext.getBeanNamesForType(BitcoinTxRepository::class.java).map { beanName ->
-            val chainName = beanName.substringBefore(REPOSITORY_NAME_DELIMETER)
+            val chainName = beanName.substringBefore(REPOSITORY_NAME_DELIMITER)
 
             val txRepository = applicationContext.getBean(beanName, BitcoinTxRepository::class.java)
